requirements.txt
psutil>=5.9.0
falkordb
fastapi
groq
jellyfish
langchain-core
langchain-groq
leidenalg
llama-index
mcp
networkx
nltk
numpy
pandas
psutil
pydantic
pydantic-ai
pykeen
python-dotenv
python-igraph
pyyaml
scikit-learn
sentence-transformers
torch
uvicorn
llama-index-embeddings-huggingface
llama-index-llms-groq
psycopg2-binary
pgvector